test non-2d in AnnData.unwriteable
