+++ tests/test_readwrite.py
@@ -174,6 +174,28 @@ def test_can_not_write_with_custom_array(
     assert adata.unwriteable(store_type=store_type)
 
 
+@pytest.mark.parametrize("store_type", ["h5", "zarr", None])
+def test_unwriteable_non_2d(
+    arr2d: np.ndarray,
+    arr3d: np.ndarray,
+    which: Literal["X", "layers"],
+    store_type: Literal["h5", "zarr"] | None,
+) -> None:
+
+    def _set_non_2d(
+        adata: ad.AnnData, which: Literal["X", "layers"], value: np.ndarray
+    ) -> None:
+        if which == "X":
+            adata.X = value
+        else:
+            adata.layers["L"] = value
+
+    adata = ad.AnnData(X=arr2d)
+    with pytest.warns(UserWarning, match=r"must be 2-dimensional"):
+        _set_non_2d(adata, which, arr3d)
+    assert adata.unwriteable(store_type=store_type)
+
+
 @pytest.mark.parametrize("typ", ARRAY_TYPES)
 def test_readwrite_roundtrip(typ, tmp_path, diskfmt, diskfmt2):
     pth1 = tmp_path / f"first.{diskfmt}"
